pyproject.toml
[tool.poetry]
name = "fin-manager"
version = "0.1.0"
description = ""
authors = ["Nikita Meschaninov <meshchaninov.n@gmail.com>", "Vladimir Chichkin <vovacickin3386@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
django-environ = "^0.8.1"
django-dotenv = "^1.4.2"
Markdown = "^3.3.6"
django-filter = "^21.1"
djangorestframework-simplejwt = "^5.1.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
django-types = "^0.15.0"
djangorestframework-types = "^0.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "fin_manager.settings.dev"
